Moved all the OpenGL 3 changes into a separate implementation called gl3fontstash.h
